backport netdev_notifier_info_to_dev()
this is a bit tricky - newer kernels pass netdev_notifier_info
to the notifier, while older kernels pass the net_device itself.

instead of adding patching the callers, add a dummy function
to return the dev itself.

don't use LINUX_BACKPORT, so the build will break if this
struct/function is already defined.

Signed-off-by: Eliad Peller <eliad@wizery.com>